pyproject.toml
[project]
name = "django-accsybiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==4.2.20",
    "django-autotyping>=0.5.1",
    "django-ckeditor-5>=0.2.18",
    "django-cotton>=2.1.3",
    "django-stubs>=5.2.2",
    "django-vote>=2.4.0",
    "polars>=1.33.0",
]


[tool.pyright]
reportIncompatibleVariableOverride = false
autoImportCompletions = true
